Refactor MatchesZip to MatchesFile
